refactor: 提取 OCR 方法到 3 个 partial class 文件
- FmMain.OCR.Core.cs: 截图→OCR→显示 主流程 (4 methods)
- FmMain.OCR.Engines.cs: 各 OCR 引擎实现 (14 methods)
- FmMain.OCR.MenuHandlers.cs: OCR 菜单处理器 (21 methods)
