+++ FmMain.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Drawing;
@@ -1335,579 +1335,7 @@ namespace TrOCR
 		// - OCR_youdao(): 调用有道 OCR API 进行文字识别。
 		// ====================================================================================================================
 		#region OCR 引擎实现
-		/// <summary>
-		/// 使用腾讯云OCR服务识别屏幕截图中的文本内容
-		/// </summary>
-		public void OCR_Tencent()
-		{
-			Image imageToProcess = image_screen;
-			Image tempBitmap = null;
-
-			try
-			{
-				split_txt = "";
-				typeset_txt = "";
-
-				// 判断是否使用高精度模式
-				bool isAccurate = (interface_flag == "腾讯-高精度");
-				string secretId = isAccurate ? StaticValue.TX_ACCURATE_API_ID : StaticValue.TX_API_ID;
-				string secretKey = isAccurate ? StaticValue.TX_ACCURATE_API_KEY : StaticValue.TX_API_KEY;
-				string language = isAccurate ? StaticValue.TX_ACCURATE_LANGUAGE : StaticValue.TX_LANGUAGE;
-				string apiType = isAccurate ? "GeneralAccurateOCR" : "GeneralBasicOCR";
-
-				// 检查密钥是否已配置
-				if (string.IsNullOrEmpty(secretId) || string.IsNullOrEmpty(secretKey))
-				{
-					typeset_txt = isAccurate ? "***请在设置中输入腾讯云高精度版密钥***" : "***请在设置中输入腾讯云密钥***";
-					split_txt = typeset_txt;
-					return;
-				}
-
-				// 调整图像尺寸以适应OCR识别要求
-				if (imageToProcess.Width > 90 && imageToProcess.Height < 90)
-				{
-					tempBitmap = new Bitmap(imageToProcess.Width, 300);
-					using (Graphics graphics = Graphics.FromImage(tempBitmap))
-					{
-						graphics.DrawImage(imageToProcess, 5, 0, imageToProcess.Width, imageToProcess.Height);
-					}
-					imageToProcess = tempBitmap;
-				}
-				else if (imageToProcess.Width <= 90 && imageToProcess.Height >= 90)
-				{
-					tempBitmap = new Bitmap(300, imageToProcess.Height);
-					using (Graphics graphics2 = Graphics.FromImage(tempBitmap))
-					{
-						graphics2.DrawImage(imageToProcess, 0, 5, imageToProcess.Width, imageToProcess.Height);
-					}
-					imageToProcess = tempBitmap;
-				}
-				else if (imageToProcess.Width < 90 && imageToProcess.Height < 90)
-				{
-					tempBitmap = new Bitmap(300, 300);
-					using (Graphics graphics3 = Graphics.FromImage(tempBitmap))
-					{
-						graphics3.DrawImage(imageToProcess, 5, 5, imageToProcess.Width, imageToProcess.Height);
-					}
-					imageToProcess = tempBitmap;
-				}
-
-				// 将图像转换为字节数组并调用腾讯OCR接口
-				byte[] imageBytes = OcrHelper.ImgToBytes(imageToProcess);
-
-				string result = TencentOcrHelper.Ocr(imageBytes, secretId, secretKey, apiType, language);
-				typeset_txt = result;
-				split_txt = result;
-			}
-			catch (Exception ex)
-			{
-				typeset_txt = $"***腾讯OCR识别出错: {ex.Message}***";
-				split_txt = typeset_txt;
-				if (esc == "退出")
-				{
-					esc = "";
-				}
-			}
-			finally
-			{
-				tempBitmap?.Dispose();
-			}
-		}
-
-		/// <summary>
-		/// 使用微信OCR服务识别屏幕截图中的文本内容
-		/// </summary>
-		public void OCR_WeChat()
-		{
-			try
-			{
-				split_txt = "";
-				typeset_txt = "";
-				// 将图像转换为字节数组并调用微信OCR接口
-				byte[] imageBytes = OcrHelper.ImgToBytes(image_screen);
-				string result = OcrHelper.WeChat(imageBytes).GetAwaiter().GetResult();
-				typeset_txt = result;
-				split_txt = result;
-			}
-			catch (Exception ex)
-			{
-				if (RuntimeInformation.ProcessArchitecture != Architecture.X64)
-				{
-					typeset_txt = split_txt="***微信OCR仅支持64位系统,不支持32位系统***";
-					return;
-					
-				}
-				typeset_txt = $"***微信OCR识别出错: {ex.Message}***";
-				if (esc == "退出")
-				{
-					esc = "";
-				}
-			}
-		}
-
-		/// <summary>
-		/// 使用白描OCR服务识别屏幕截图中的文本内容
-		/// </summary>
-		public void OCR_Baimiao()
-		{
-			try
-			{
-				split_txt = "";
-				typeset_txt = "";
-				
-				// 将图像转换为字节数组并调用白描OCR接口
-				byte[] imageBytes = OcrHelper.ImgToBytes(image_screen);
-				// 调用已重构的、无参数的Baimiao方法
-				string result = OcrHelper.Baimiao(imageBytes).GetAwaiter().GetResult();
-				typeset_txt = result;
-				split_txt = result;
-			}
-			catch (Exception ex)
-			{
-				typeset_txt = $"***白描OCR识别出错: {ex.Message}***";
-				split_txt = typeset_txt;
-				if (esc == "退出")
-				{
-					esc = "";
-				}
-			}
-		}
-
-		/// <summary>
-		/// 使用百度OCR服务识别屏幕截图中的文本内容
-		/// 调用百度OCR通用文字识别API进行文字识别，并根据识别结果更新文本框内容
-		/// </summary>
-		public void OCR_baidu()
-		{
-			split_txt = "";
-			try
-			{
-		  				// 从 StaticValue 读取语言类型
-		  				string languageType = StaticValue.BD_LANGUAGE;
-
-		  var imageBytes = OcrHelper.ImgToBytes(image_screen);
-		  // 调用已重构的、无密钥参数的方法
-		  var result = BaiduOcrHelper.GeneralBasic(imageBytes, languageType);
-
-		  if (!string.IsNullOrEmpty(result))
-		  {
-					if (result.StartsWith("***") || result.Contains("错误") || result.Contains("失败"))
-					{
-						// 错误信息直接显示
-						if (esc != "退出")
-						{
-							RichBoxBody.Text = result;
-						}
-						else
-						{
-							RichBoxBody.Text = "***该区域未发现文本***";
-							esc = "";
-						}
-					}
-					else
-					{
-						// 处理识别结果
-						ProcessOcrResult(result);
-					}
-				}
-				else
-				{
-					RichBoxBody.Text = "***百度OCR识别失败***";
-				}
-			}
-			catch (Exception ex)
-			{
-				if (esc != "退出")
-				{
-					RichBoxBody.Text = "***该区域未发现文本或者密钥次数用尽***";
-				}
-				else
-				{
-					RichBoxBody.Text = "***该区域未发现文本***";
-					esc = "";
-				}
-			}
-		}
-
-		/// <summary>
-		/// 百度OCR高精度版
-		/// 使用百度OCR高精度版服务识别屏幕截图中的文本内容
-		/// </summary>
-		public void OCR_baidu_accurate()
-		{
-			split_txt = "";
-			try
-			{
-		              // 从 StaticValue 读取高精度版设置
-		              string languageType = StaticValue.BD_ACCURATE_LANGUAGE;
-
-		  var imageBytes = OcrHelper.ImgToBytes(image_screen);
-		  // 调用已重构的、无密钥参数的方法
-		  var result = BaiduOcrHelper.AccurateBasic(imageBytes, languageType);
-
-		  if (!string.IsNullOrEmpty(result))
-		  {
-					if (result.StartsWith("***") || result.Contains("错误") || result.Contains("失败"))
-					{
-						// 错误信息直接显示
-						if (esc != "退出")
-						{
-							RichBoxBody.Text = result;
-						}
-						else
-						{
-							RichBoxBody.Text = "***该区域未发现文本***";
-							esc = "";
-						}
-					}
-					else
-					{
-						// 处理识别结果
-						ProcessOcrResult(result);
-					}
-				}
-				else
-				{
-					RichBoxBody.Text = "***百度高精度OCR识别失败***";
-				}
-			}
-			catch (Exception ex)
-			{
-				if (esc != "退出")
-				{
-					RichBoxBody.Text = "***该区域未发现文本或者密钥次数用尽***";
-				}
-				else
-				{
-					RichBoxBody.Text = "***该区域未发现文本***";
-					esc = "";
-				}
-			}
-		}
-
-        /// <summary>
-		/// 处理OCR识别结果
-		/// 将OCR识别出的文本结果进行处理和格式化
-        /// </summary>
-		/// <param name="result">OCR识别出的原始文本结果</param>
-        private void ProcessOcrResult(string result)
-        {
-			// 将纯文本结果转换为之前的格式进行处理
-			var lines = result.Split(new[] { "\r\n", "\n" }, StringSplitOptions.RemoveEmptyEntries);
-            var jArray = new JArray();
-            foreach (var line in lines)
-            {
-				if (!string.IsNullOrWhiteSpace(line))
-                {
-                    var jObject = new JObject();
-					jObject["words"] = line;
-                    jArray.Add(jObject);
-                }
-            }
-
-            if (jArray.Count > 0)
-            {
-                checked_txt(jArray, 1, "words");
-            }
-            else
-            {
-                split_txt = "";
-                typeset_txt = "";
-            }
-        }
-
-        /// <summary>
-        /// PaddleOCR离线识别方法
-        /// 使用PaddleOCR引擎进行本地离线文字识别
-        /// </summary>
-        public void OCR_PaddleOCR()
-		{
-			split_txt = "";
-			try
-			{
-				var result = PaddleOCRHelper.RecognizeText(image_screen);
-				image_screen?.Dispose();
-				// GC.Collect();
-				// GC.WaitForPendingFinalizers();
-
-				if (!string.IsNullOrEmpty(result))
-				{
-					if (result.StartsWith("***") || result.Contains("错误") || result.Contains("失败"))
-					{
-						// 错误信息直接显示
-						if (esc != "退出")
-						{
-							// RichBoxBody.Text = result;//这里有bug，所以改为下面两行代码
-							typeset_txt = result;
-							split_txt = typeset_txt; // 必须也把这个变量设置一下
-						}
-						else
-						{
-							typeset_txt = "***该区域未发现文本***";
-							split_txt = typeset_txt;
-							esc = "";
-						}
-					}
-					else
-					{
-                        // 处理识别结果
-                        //ProcessOcrResult(result);
-                        // 匹配 1次 或 多次 连续的换行
-                        typeset_txt = Regex.Replace(result, @"(\r\n)+", "\r\n");
-                        split_txt = typeset_txt;
-                    }
-				}
-				else
-				{ //这里应该也要改，先标记一下，暂时不改
-					RichBoxBody.Text = "***PaddleOCR识别失败***";
-				}
-			}
-			catch (Exception ex)
-			{
-				//if (RuntimeInformation.ProcessArchitecture != Architecture.X64)
-				//{
-				//    typeset_txt = split_txt = "***PaddleOCR仅支持64位系统,不支持32位系统***";
-				//    return;
-
-				//}
-				if (esc != "退出")
-				{
-					//RichBoxBody.Text = "***PaddleOCR识别失败: " + ex.Message + "***";这里有bug，所以改为下面两行代码
-					typeset_txt = "***PaddleOCR识别失败: " + ex.Message + "***";
-					split_txt = typeset_txt; // 必须也把这个变量设置一下
-				}
-				else
-				{
-					typeset_txt = "***该区域未发现文本***";
-					split_txt = typeset_txt;
-					esc = "";
-				}
-            }
-			 TrOCRUtils.CleanMemory();
-		}
-
-		/// <summary>
-		/// PaddleOCR2离线识别方法
-		/// 使用Sdcb.PaddleOCR引擎进行本地离线文字识别
-		/// </summary>
-		public void OCR_PaddleOCR2()
-		{
-			split_txt = "";
-			try
-			{
-				var result = PaddleOCR2Helper.RecognizeText(image_screen);
-				image_screen?.Dispose();
-                // GC.Collect();
-                // GC.WaitForPendingFinalizers();
-
-                if (!string.IsNullOrEmpty(result))
-				{
-					if (result.StartsWith("***") || result.Contains("错误") || result.Contains("失败"))
-					{
-						// 错误信息直接显示
-						if (esc != "退出")
-						{
-							typeset_txt = result;
-                    		split_txt = typeset_txt;
-						}
-						else
-						{
-							typeset_txt = "***该区域未发现文本***";
-							split_txt = typeset_txt;
-							esc = "";
-						}
-					}
-					else
-					{
-                        // 处理识别结果
-                        //ProcessOcrResult(result);
-                        // 匹配 1次 或 多次 连续的换行
-                        typeset_txt = Regex.Replace(result, @"(\r\n)+", "\r\n");
-                        split_txt = typeset_txt;
-                    }
-				}
-				else
-				{
-					RichBoxBody.Text = "***PaddleOCR2识别失败***";
-				}
-            }
-			catch (Exception ex)
-			{
-                //if (RuntimeInformation.ProcessArchitecture != Architecture.X64)
-                //{
-                //    typeset_txt = split_txt = "***PaddleOCR2仅支持64位系统,不支持32位系统***";
-                //    return;
-
-                //}
-                if (esc != "退出")
-				{
-                    typeset_txt = "***PaddleOCR2识别失败: " + ex.Message + "***";
-                    split_txt = typeset_txt;
-                }
-				else
-				{
-                    typeset_txt = "***该区域未发现文本***";
-                    split_txt = typeset_txt;
-                    esc = "";
-                }
-            }
-			 TrOCRUtils.CleanMemory();
-		}
-
-		/// <summary>
-		/// RapidOCR离线识别方法
-		/// 使用RapidOCR引擎进行本地离线文字识别
-		/// </summary>
-		public void OCR_RapidOCR()
-		{
-			split_txt = "";
-			try
-			{
-				var result = RapidOCRHelper.RecognizeText(image_screen);
-				
-				if (!string.IsNullOrEmpty(result))
-				{
-					if (result.StartsWith("***") || result.Contains("错误") || result.Contains("失败"))
-					{
-						// 错误信息直接显示
-						if (esc != "退出")
-						{
-							typeset_txt = result;
-                   		split_txt = typeset_txt; // 必须也把这个变量设置一下
-						}
-						else
-						{
-							typeset_txt = "***该区域未发现文本***";
-							split_txt = typeset_txt;
-							esc = "";
-						}
-					}
-					else
-					{
-                        // 处理识别结果
-                        //ProcessOcrResult(result);
-                        // 匹配 1次 或 多次 连续的换行
-                        typeset_txt = Regex.Replace(result, @"(\r\n)+", "\r\n");
-                        split_txt = typeset_txt;
-                    }
-				}
-				else
-				{
-					RichBoxBody.Text = "***RapidOCR识别失败***";
-				}
-            }
-			catch (Exception ex)
-			{
-				if (esc != "退出")
-				{
-                   typeset_txt = "***RapidOCR识别失败: " + ex.Message + "***";
-                   split_txt = typeset_txt; // 必须也把这个变量设置一下
-               }
-				else
-				{
-                   typeset_txt = "***该区域未发现文本***";
-                   split_txt = typeset_txt;
-                   esc = "";
-               }
-            }
-			 TrOCRUtils.CleanMemory();
-		}
 		#endregion
-
-		/// <summary>
-		/// 有道OCR识别方法
-		/// 调用有道OCR接口进行文字识别，对图像进行预处理以提高识别准确率
-		/// </summary>
-		public void OCR_youdao()
-		{
-			try
-			{
-				split_txt = "";
-				var image = image_screen;
-				// 对过小的图像进行填充以达到合适的识别尺寸
-				if (image.Width > 90 && image.Height < 90)
-				{
-					var bitmap = new Bitmap(image.Width, 200);
-					var graphics = Graphics.FromImage(bitmap);
-					graphics.DrawImage(image, 5, 0, image.Width, image.Height);
-					graphics.Save();
-					graphics.Dispose();
-					image = new Bitmap(bitmap);
-				}
-				else if (image.Width <= 90 && image.Height >= 90)
-				{
-					var bitmap2 = new Bitmap(200, image.Height);
-					var graphics2 = Graphics.FromImage(bitmap2);
-					graphics2.DrawImage(image, 0, 5, image.Width, image.Height);
-					graphics2.Save();
-					graphics2.Dispose();
-					image = new Bitmap(bitmap2);
-				}
-				else if (image.Width < 90 && image.Height < 90)
-				{
-					var bitmap3 = new Bitmap(200, 200);
-					var graphics3 = Graphics.FromImage(bitmap3);
-					graphics3.DrawImage(image, 5, 5, image.Width, image.Height);
-					graphics3.Save();
-					graphics3.Dispose();
-					image = new Bitmap(bitmap3);
-				}
-				else
-				{
-					image = image_screen;
-				}
-				var i = image.Width;
-				var j = image.Height;
-				// 对图像进行放大处理以提高识别准确率
-				if (i < 600)
-				{
-					while (i < 600)
-					{
-						j *= 2;
-						i *= 2;
-					}
-				}
-				if (j < 120)
-				{
-					while (j < 120)
-					{
-						j *= 2;
-						i *= 2;
-					}
-				}
-				var bitmap4 = new Bitmap(i, j);
-				var graphics4 = Graphics.FromImage(bitmap4);
-				graphics4.DrawImage(image, 0, 0, i, j);
-				graphics4.Save();
-				graphics4.Dispose();
-				image = new Bitmap(bitmap4);
-				var inArray = OcrHelper.ImgToBytes(image);
-				var data = "imgBase=data" + HttpUtility.UrlEncode(":image/jpeg;base64," + Convert.ToBase64String(inArray)) + "&lang=auto&company=";
-				var value = CommonHelper.PostStrData("http://aidemo.youdao.com/ocrapi1", data, "",
-					"http://aidemo.youdao.com/ocrdemo");
-				var jArray = JArray.Parse(((JObject)JsonConvert.DeserializeObject(value))["lines"].ToString());
-				checked_txt(jArray, 1, "words");
-				image.Dispose();
-			}
-			catch
-			{
-				if (esc != "退出")
-				{
-					if (RichBoxBody.Text != "***该区域未发现文本***")
-					{
-						RichBoxBody.Text = "***该区域未发现文本***";
-					}
-				}
-				else
-				{
-					if (RichBoxBody.Text != "***该区域未发现文本***")
-					{
-						RichBoxBody.Text = "***该区域未发现文本***";
-					}
-					esc = "";
-				}
-			}
-		}
 
 // ====================================================================================================================
 		// **OCR 接口切换 (事件)**
@@ -1916,123 +1344,6 @@ namespace TrOCR
 		// 每个事件处理程序通过调用 OCR_foreach(string name) 方法来更新当前使用的 OCR 接口。
 		// ====================================================================================================================
 		#region OCR 接口切换 (事件)
-		/// <summary>
-		/// 搜狗OCR接口选择事件处理函数
-		/// 切换当前OCR接口为搜狗OCR
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_sougou_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("搜狗");
-		}
-
-		/// <summary>
-		/// 腾讯OCR接口选择事件处理函数
-		/// 切换当前OCR接口为腾讯OCR
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_tencent_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("腾讯");
-		}
-
-		/// <summary>
-		/// 腾讯高精度OCR接口选择事件处理函数
-		/// 切换当前OCR接口为腾讯高精度OCR
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_tencent_accurate_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("腾讯-高精度");
-		}
-
-		/// <summary>
-		/// 百度OCR接口选择事件处理函数
-		/// 切换当前OCR接口为百度OCR
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_baidu_Click(object sender, EventArgs e)
-		{
-		}
-
-		/// <summary>
-		/// 有道OCR接口选择事件处理函数
-		/// 切换当前OCR接口为有道OCR
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void OCR_youdao_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("有道");
-		}
-
-		/// <summary>
-		/// 微信OCR接口选择事件处理函数
-		/// 切换当前OCR接口为微信OCR
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void OCR_wechat_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("微信");
-		}
-
-		/// <summary>
-		/// 白描OCR接口选择事件处理函数
-		/// 切换当前OCR接口为白描OCR
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		public void OCR_baimiao_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("白描");
-		}
-
-		/// <summary>
-		/// 百度高精度OCR接口选择事件处理函数
-		/// 切换当前OCR接口为百度高精度OCR
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_baidu_accurate_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("百度-高精度");
-		}
-
-		/// <summary>
-		/// PaddleOCR菜单点击事件处理
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_paddleocr_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("PaddleOCR");
-		}
-
-		/// <summary>
-		/// PaddleOCR2菜单点击事件处理
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_paddleocr2_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("PaddleOCR2");
-		}
-
-		/// <summary>
-		/// RapidOCR菜单点击事件处理
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_rapidocr_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("RapidOCR");
-		}
-
 		#endregion
 // ====================================================================================================================
 		// **文本操作与格式化**
@@ -4083,936 +3394,6 @@ namespace TrOCR
 		// ====================================================================================================================
 #region 截图与核心OCR流程
 		/// <summary>
-		/// 主OCR快速截图功能
-		/// 启动截图功能，隐藏主窗口，调用ShareX库进行区域捕捉
-		/// 根据用户的操作（如截图、贴图、保存、多区域选择等）执行不同逻辑
-		/// </summary>
-		public void MainOCRQuickScreenShots()
-		{
-			// 如果正在截图则直接返回
-			if (StaticValue.IsCapture) return;
-            // 【新增】进入截图模式前，停止所有翻译定时器
-            if (translationTimer != null) translationTimer.Stop();
-            try
-			{
-				// 隐藏主窗口并准备截图
-				change_QQ_screenshot = false;
-
-				// === 纵深防御: 挂起 lastNormalSize 自动更新 ===
-				isProgrammaticResize = true;
-
-				// === 先关闭翻译模式，再操作 FormBorderStyle ===
-				transtalate_fla = "关闭";
-
-				// 如果工具栏翻译功能关闭，则执行关闭翻译操作
-				if (IniHelper.GetValue("工具栏", "翻译") == "False")
-				{
-					Trans_close_Click(null, EventArgs.Empty, false);
-				}
-
-				// === 现在安全地改变 FormBorderStyle ===
-				FormBorderStyle = FormBorderStyle.None;
-				Visible = false;
-				Thread.Sleep(100);
-
-				form_width = Width;
-				
-				// 初始化相关变量
-				shupai_Right_txt = "";
-				shupai_Left_txt = "";
-				form_height = Height;
-				// === 恢复 lastNormalSize 自动更新 ===
-				isProgrammaticResize = false;
-				minico.Visible = false;
-				minico.Visible = true;
-				menu.Close();
-				menu_copy.Close();
-				auto_fla = "开启";
-				split_txt = "";
-
-				// --- 步骤 1: 暂时断开事件处理程序 ---
-				RichBoxBody.richTextBox1.TextChanged -= RichBoxBody_TextChanged;
-
-				try
-				{
-				    // --- 步骤 2: 执行“静默”更新 ---
-				    // 避免不必要的文本更新
-				    if (RichBoxBody.Text != "***该区域未发现文本***")
-				    {
-				        RichBoxBody.Text = "***该区域未发现文本***";
-				    }
-				    RichBoxBody_T.Text = "";
-				    typeset_txt = "";
-				}
-				finally
-				{
-				    // 没必要重新绑定了
-				    //RichBoxBody.richTextBox1.TextChanged += RichBoxBody_TextChanged;
-				}
-				
-				// 重置窗口大小和边框样式
-				// Size = new Size((int)font_base.Width * 23, (int)font_base.Height * 24);
-				FormBorderStyle = FormBorderStyle.Sizable;
-				
-				// 设置截图状态为进行中
-				StaticValue.IsCapture = true;
-				image_screen?.Dispose();
-				// 调用截图功能获取屏幕图像
-				image_screen = RegionCaptureTasks.GetRegionImage_Mo(new RegionCaptureOptions
-				{
-					ShowMagnifier = false,
-					UseSquareMagnifier = false,
-					MagnifierPixelCount = 15,
-					MagnifierPixelSize = 10
-				}, out var modeFlag, out var point, out var buildRects);
-
-				// 如果是静默模式，强制进行OCR，忽略截图工具栏的其他按钮功能
-				if (isSilentMode && image_screen != null)
-				{
-				    modeFlag = "SilentOcrTrigger"; // 使用一个不存在的标志来触发switch case默认的OCR流程
-				}
-				
-				// 如果是高级截图模式，则启动高级截图窗体
-				if (modeFlag == "高级截图")
-				{
-					var mode = RegionCaptureMode.Annotation;
-					var options = new RegionCaptureOptions();
-					using (var regionCaptureForm = new RegionCaptureForm(mode, options))
-					{
-						regionCaptureForm.Image_get = false;
-						regionCaptureForm.Prepare(image_screen);
-						regionCaptureForm.ShowDialog();
-						image_screen = null;
-						image_screen = regionCaptureForm.GetResultImage();
-						modeFlag = regionCaptureForm.Mode_flag;
-					}
-				}
-				
-				// 注册ESC键作为退出截图的热键
-				HelpWin32.RegisterHotKey(Handle, 222, HelpWin32.KeyModifiers.None, Keys.Escape);
-				
-				// 根据截图后的操作模式执行相应处理
-				switch (modeFlag)
-				{
-					case "贴图":
-						{
-							// 贴图模式：创建贴图窗体并显示
-							var locationPoint = new Point(point.X, point.Y);
-							new FmScreenPaste(image_screen, locationPoint).Show();
-							if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-							{
-								var value = IniHelper.GetValue("快捷键", "翻译文本");
-								var text = "None";
-								var text2 = "F9";
-								SetHotkey(text, text2, value, 205);
-							}
-							HelpWin32.UnregisterHotKey(Handle, 222);
-							StaticValue.IsCapture = false;
-							break;
-						}
-					case "区域多选" when image_screen == null:
-						{
-							// 区域多选但未选择区域：恢复热键并退出截图状态
-							if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-							{
-								var value2 = IniHelper.GetValue("快捷键", "翻译文本");
-								var text3 = "None";
-								var text4 = "F9";
-								SetHotkey(text3, text4, value2, 205);
-							}
-							HelpWin32.UnregisterHotKey(Handle, 222);
-							StaticValue.IsCapture = false;
-							break;
-						}
-					case "区域多选":
-						// 区域多选：启动加载线程并处理多个区域的OCR
-						minico.Visible = true;
-						thread = new Thread(ShowLoading);
-						thread.Start();
-						ts = new TimeSpan(DateTime.Now.Ticks);
-						getSubPics_ocr(image_screen, buildRects);
-						break;
-					case "取色":
-						{
-							// 取色模式：恢复热键并显示颜色已复制提示
-							if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-							{
-								var value3 = IniHelper.GetValue("快捷键", "翻译文本");
-								var text5 = "None";
-								var text6 = "F9";
-								SetHotkey(text5, text6, value3, 205);
-							}
-							HelpWin32.UnregisterHotKey(Handle, 222);
-							StaticValue.IsCapture = false;
-							CommonHelper.ShowHelpMsg("已复制颜色");
-							break;
-						}
-					default:
-						{
-							if (image_screen == null)
-							{
-								// 未获取到图像：恢复热键并退出截图状态
-								if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-								{
-									var value4 = IniHelper.GetValue("快捷键", "翻译文本");
-									var text7 = "None";
-									var text8 = "F9";
-									SetHotkey(text7, text8, value4, 205);
-								}
-								HelpWin32.UnregisterHotKey(Handle, 222);
-								StaticValue.IsCapture = false;
-							}
-							else
-							{
-								// 根据不同模式标志设置相应变量
-								if (modeFlag == "百度")
-								{
-									baidu_flags = "百度";
-								}
-								if (modeFlag == "拆分")
-								{
-									set_merge = false;
-									set_split = true;
-								}
-								if (modeFlag == "合并")
-								{
-									set_merge = true;
-									set_split = false;
-								}
-								if (modeFlag == "截图")
-								{
-									// 截图模式：将图像复制到剪贴板
-									Clipboard.SetImage(image_screen);
-									if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-									{
-										var value5 = IniHelper.GetValue("快捷键", "翻译文本");
-										var text9 = "None";
-										var text10 = "F9";
-										SetHotkey(text9, text10, value5, 205);
-									}
-									HelpWin32.UnregisterHotKey(Handle, 222);
-									StaticValue.IsCapture = false;
-									if (IniHelper.GetValue("截图音效", "粘贴板") == "True")
-									{
-										PlaySong(IniHelper.GetValue("截图音效", "音效路径"));
-									}
-									CommonHelper.ShowHelpMsg("已复制截图");
-								}
-								else if (modeFlag == "自动保存" && IniHelper.GetValue("配置", "自动保存") == "True")
-								{
-									// 自动保存模式：将图像保存到指定位置
-									var filename = IniHelper.GetValue("配置", "截图位置") + "\\" + ReFileName(IniHelper.GetValue("配置", "截图位置"), "图片.Png");
-									image_screen.Save(filename, ImageFormat.Png);
-									StaticValue.IsCapture = false;
-									if (IniHelper.GetValue("截图音效", "自动保存") == "True")
-									{
-										PlaySong(IniHelper.GetValue("截图音效", "音效路径"));
-									}
-									CommonHelper.ShowHelpMsg("已保存图片");
-								}
-								else if (modeFlag == "多区域自动保存" && IniHelper.GetValue("配置", "自动保存") == "True")
-								{
-									// 多区域自动保存模式：保存多个区域的图像
-									getSubPics(image_screen, buildRects);
-									StaticValue.IsCapture = false;
-									if (IniHelper.GetValue("截图音效", "自动保存") == "True")
-									{
-										PlaySong(IniHelper.GetValue("截图音效", "音效路径"));
-									}
-									CommonHelper.ShowHelpMsg("已保存图片");
-								}
-								else if (modeFlag == "保存")
-								{
-									// 保存模式：弹出保存对话框让用户选择保存位置和格式
-									var saveFileDialog = new SaveFileDialog();
-									saveFileDialog.Filter = "png图片(*.png)|*.png|jpg图片(*.jpg)|*.jpg|bmp图片(*.bmp)|*.bmp";
-									saveFileDialog.AddExtension = false;
-									saveFileDialog.FileName = string.Concat("tianruo_", DateTime.Now.Year.ToString(), "-", DateTime.Now.Month.ToString(), "-", DateTime.Now.Day.ToString(), "-", DateTime.Now.Ticks.ToString());
-									saveFileDialog.Title = "保存图片";
-									saveFileDialog.FilterIndex = 1;
-									saveFileDialog.RestoreDirectory = true;
-									if (saveFileDialog.ShowDialog() == DialogResult.OK)
-									{
-										var extension = Path.GetExtension(saveFileDialog.FileName);
-										if (extension.Equals(".jpg"))
-										{
-											image_screen.Save(saveFileDialog.FileName, ImageFormat.Jpeg);
-										}
-										if (extension.Equals(".png"))
-										{
-											image_screen.Save(saveFileDialog.FileName, ImageFormat.Png);
-										}
-										if (extension.Equals(".bmp"))
-										{
-											image_screen.Save(saveFileDialog.FileName, ImageFormat.Bmp);
-										}
-									}
-									if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-									{
-										var value6 = IniHelper.GetValue("快捷键", "翻译文本");
-										var text11 = "None";
-										var text12 = "F9";
-										SetHotkey(text11, text12, value6, 205);
-									}
-									HelpWin32.UnregisterHotKey(Handle, 222);
-									StaticValue.IsCapture = false;
-								}
-								else if (image_screen != null)
-								{
-									// OCR识别模式：根据工具栏设置决定是否进行分栏处理
-									if (IniHelper.GetValue("工具栏", "分栏") == "True")
-									{
-										minico.Visible = true;
-										thread = new Thread(ShowLoading);
-										thread.Start();
-										ts = new TimeSpan(DateTime.Now.Ticks);
-										var image = image_screen;
-										var graphics = Graphics.FromImage(new Bitmap(image.Width, image.Height));
-										graphics.DrawImage(image, 0, 0, image.Width, image.Height);
-										graphics.Save();
-										graphics.Dispose();
-										((Bitmap)FindBoundingBoxFences((Bitmap)image)).Save("Data\\分栏预览图.jpg");
-										image.Dispose();
-										image_screen.Dispose();
-									}
-									else
-									{
-										// 启动OCR识别线程
-										minico.Visible = true;
-										thread = new Thread(ShowLoading);
-										thread.Start();
-										ts = new TimeSpan(DateTime.Now.Ticks);
-										var messageload = new Messageload();
-										messageload.ShowDialog();
-										if (messageload.DialogResult == DialogResult.OK)
-										{
-											esc_thread = new Thread(Main_OCR_Thread);
-											esc_thread.Start();
-										}
-									}
-								}
-							}
-
-							break;
-						}
-				}
-			}
-			catch
-			{
-				// 发生异常时确保退出截图状态
-				StaticValue.IsCapture = false;
-			}
-		}
-
-		/// <summary>
-		/// OCR主线程函数，根据不同的接口标识调用相应的OCR识别方法，并处理识别结果
-		/// </summary>
-		public void Main_OCR_Thread()
-		{
-			// 优先检查是否为二维码，如果是则直接返回二维码内容
-			string qrCodeResult = ScanQRCode();
-			if (!string.IsNullOrEmpty(qrCodeResult))
-			{
-				typeset_txt = qrCodeResult;
-                split_txt = qrCodeResult;
-                RichBoxBody.Text = typeset_txt;
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			// 根据interface_flag选择不同的OCR接口进行识别
-			if (interface_flag == "搜狗")
-			{
-				SougouOCR();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			if (interface_flag == "腾讯" || interface_flag == "腾讯-高精度")
-			{
-				OCR_Tencent();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			if (interface_flag == "有道")
-			{
-				OCR_youdao();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			if (interface_flag == "微信")
-			{
-				OCR_WeChat();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			if (interface_flag == "白描")
-			{
-				OCR_Baimiao();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			if (interface_flag == "公式")
-			{
-				OCR_Math();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			if (interface_flag == "百度表格")
-			{
-				BdTableOCR();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_table));
-				return;
-			}
-			if (interface_flag == "腾讯表格")
-			{
-				TxTableOCR();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_table));
-				return;
-			}
-			if (interface_flag == "阿里表格")
-			{
-				OCR_ali_table();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_table));
-				return;
-			}
-			if (interface_flag == "日语" || interface_flag == "中英" || interface_flag == "韩语")
-			{
-				OCR_baidu();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-			}
-			if (interface_flag == "百度-高精度")
-			{
-				OCR_baidu_accurate();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-			}
-			if (interface_flag == "CustomOpenAI")
-			{
-				// 调用 FmMain.AI.cs 里的执行方法
-				// OCR_Custom_Router(); 
-				OCR_OpenAICompatible();
-				
-				// 善后工作 (关闭加载窗，显示主窗)
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			// 【新增】百度手写识别的分支
-			if (interface_flag == "百度手写")
-			{
-			    OCR_baidu_handwriting();
-			    fmloading.FmlClose = "窗体已关闭";
-			    Invoke(new OcrThread(Main_OCR_Thread_last));
-			    return;
-			}
-			if (interface_flag == "PaddleOCR")
-			{
-				OCR_PaddleOCR();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			if (interface_flag == "PaddleOCR2")
-			{
-				OCR_PaddleOCR2();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			if (interface_flag == "RapidOCR")
-			{
-				OCR_RapidOCR();
-				fmloading.FmlClose = "窗体已关闭";
-				Invoke(new OcrThread(Main_OCR_Thread_last));
-				return;
-			}
-			// 处理竖排文字识别（从左向右或从右向左）
-			if (interface_flag == "从左向右" || interface_flag == "从右向左")
-			{
-				shupai_Right_txt = "";
-				var image = image_screen;
-				var bitmap = new Bitmap(image.Width, image.Height);
-				var graphics = Graphics.FromImage(bitmap);
-				graphics.DrawImage(image, 0, 0, image.Width, image.Height);
-				graphics.Save();
-				graphics.Dispose();
-				image_ori = bitmap;
-				var image2 = bitmap.ToImage<Gray, byte>();
-                var image3 = ((Bitmap)FindBundingBox(image2.ToBitmap())).ToImage<Gray, byte>();
-                var draw = image3.Convert<Bgr, byte>();
-				var image4 = image3.Clone();
-				CvInvoke.Canny(image3, image4, 0.0, 0.0, 5, true);
-				select_image(image4, draw);
-				bitmap.Dispose();
-				image2.Dispose();
-				image3.Dispose();
-			}
-			image_screen.Dispose();
-			GC.Collect();
-		}
-
-		/// <summary>
-		/// OCR识别完成后的处理函数，负责处理识别结果、格式化文本、更新界面和执行后续操作
-		/// </summary>
-		public async void Main_OCR_Thread_last()
-		{
-			try
-			{
-				LogState("Main_OCR_Thread_last Start");
-				// --- 新增的静默模式处理逻辑 ---
-				if (isSilentMode)
-				{
-					isSilentMode = false; // 为下一次操作重置标志
-
-					// 检查识别是否成功
-					bool success = typeset_txt != null &&
-								   !typeset_txt.Contains("***该区域未发现文本***") &&
-								   !string.IsNullOrWhiteSpace(typeset_txt);
-
-					if (success)
-					{
-						SetClipboardWithLock(typeset_txt);
-						Debug.WriteLine("静默[OCR] 识别成功，已复制到剪贴板");
-						CommonHelper.ShowHelpMsg("已复制到剪贴板");
-					}
-					else
-					{
-						string errorMessage = string.IsNullOrWhiteSpace(typeset_txt) ? "未识别到文本" : typeset_txt.Replace("***", "").Trim();
-						CommonHelper.ShowHelpMsg("静默识别失败：" + errorMessage);
-					}
-
-					HelpWin32.UnregisterHotKey(Handle, 222); // 注销ESC热键
-					StaticValue.IsCapture = false; // 确保截图状态被重置
-					image_screen?.Dispose(); // 释放图像资源
-					return; // 结束方法，不显示主窗口
-				}
-				// --- 步骤 1: 数据处理和准备 ---
-				// 关键新增：标记当前内容来源于OCR
-				isContentFromOcr = true;
-				isFromClipboardListener = false;
-				image_screen.Dispose();
-				StaticValue.IsCapture = false;
-				var text = typeset_txt;
-				// 【ai流式输出模式下，跳过旧式的标点和空格清洗，信任 AI 的原样输出】
-				if (!isStreaming)
-				{
-					text = check_str(text);
-					split_txt = check_str(split_txt);
-					// 如果文本没有标点符号，则使用拆分后的文本
-					if (!punctuation_has_punctuation(text))
-					{
-						text = split_txt;
-					}
-					// 如果包含中文，则删除空格
-					if (contain_ch(text.Trim()))
-					{
-						text = Del_Space(text);
-					}
-				}
-                StaticValue.v_Split = split_txt;
-
-                string finalTextToShow = text;
-				bool shouldPerformCopy = false;
-				string textToCopy = "";
-
-				var autoTranslate = bool.Parse(IniHelper.GetValue("工具栏", "翻译"));
-				var autoCopyOcr = StaticValue.AutoCopyOcrResult;
-				var autoCopyTranslate = StaticValue.AutoCopyOcrTranslation;
-				// 处理文本拆分选项
-				if ((bool.Parse(IniHelper.GetValue("工具栏", "拆分")) && !isStreaming) || set_split)
-				{
-					set_split = false;
-					finalTextToShow = split_txt;
-					// --- 新增: 拆分后自动复制 ---
-					if (StaticValue.IsSplitAutoCopy && !string.IsNullOrEmpty(finalTextToShow))
-					{
-						shouldPerformCopy = true;
-						textToCopy = finalTextToShow;
-					}
-				}
-				// 处理文本合并选项
-				else if ((bool.Parse(IniHelper.GetValue("工具栏", "合并")) && !isStreaming) || set_merge)
-				{
-					set_merge = false;
-					if (StaticValue.IsMergeRemoveAllSpace)
-					{
-						finalTextToShow = Regex.Replace(text, @"[\r\n 　]+", "");
-					}
-					else
-					{
-						// 只有在“非移除所有空格”模式下，才调用原来的智能合并方法
-						finalTextToShow = PerformIntelligentMerge(text, StaticValue.IsMergeRemoveSpace);
-					}
-
-				}
-
-				// 计算识别耗时
-				var timeSpan = new TimeSpan(DateTime.Now.Ticks);
-				var timeSpan2 = timeSpan.Subtract(ts).Duration();
-				var str = $"{timeSpan2.Seconds}.{Convert.ToInt32(timeSpan2.TotalMilliseconds)}秒";
-
-				// 处理笔记相关功能
-				if (finalTextToShow != null)
-				{
-					p_note(finalTextToShow);
-					StaticValue.v_note = pubnote;
-					if (fmNote.Created)
-					{
-						fmNote.TextNote = "";
-					}
-				}
-				//处理无弹窗配置
-				if (IniHelper.GetValue("配置", "识别弹窗") == "False")
-				{
-					FormBorderStyle = FormBorderStyle.Sizable;
-					// Size = new Size((int)font_base.Width * 23, (int)font_base.Height * 24);
-					this.Size = this.lastNormalSize;
-					Visible = false;
-					RichBoxBody.Text = finalTextToShow;
-					if (RichBoxBody.Text != "***该区域未发现文本***" && !string.IsNullOrWhiteSpace(RichBoxBody.Text))
-					{
-						SetClipboardWithLock(RichBoxBody.Text);
-
-						Debug.WriteLine("无弹窗模式复制识别结果成功");
-						CommonHelper.ShowHelpMsg("已识别并复制");
-					}
-					else
-					{
-						CommonHelper.ShowHelpMsg("无文本");
-					}
-					if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-					{
-						var value2 = IniHelper.GetValue("快捷键", "翻译文本");
-						var text4 = "None";
-						var text5 = "F9";
-						SetHotkey(text4, text5, value2, 205);
-					}
-					HelpWin32.UnregisterHotKey(Handle, 222);
-					return;
-				}
-
-				// --- 步骤 2: 集中进行所有UI更新 ---
-				// a. 先让窗口框架稳定
-				Text = "耗时：" + str;
-				FormBorderStyle = FormBorderStyle.Sizable;
-				// 在设置尺寸之前记录一次，这是看到 Bug 的关键
-				System.Diagnostics.Debug.WriteLine("Main_OCR_Thread_last: About to set final size.");
-				LogState("Main_OCR_Thread_last Before Set Size");
-				//Size = new Size(form_width, form_height);
-				this.Size = this.lastNormalSize;
-				// 在设置尺寸之后再记录一次
-				LogState("Main_OCR_Thread_last After Set Size");
-				if (StaticValue.v_topmost)
-				{
-					TopMost = true;
-				}
-				else
-				{
-					TopMost = false;
-				}
-				minico.Visible = true;
-				Visible = true;
-				Show();
-				WindowState = FormWindowState.Normal;
-				HelpWin32.SetForegroundWindow(Handle);
-                // ====================【核心修改开始】====================
-
-                // a. 隐藏控件，暂停渲染
-                // 非流式才隐藏，防止闪烁
-                if (!isStreaming)
-                {
-                    RichBoxBody.Visible = false;
-                }
-
-                // b. 在窗口稳定后，再填充文本内容
-                RichBoxBody.richTextBox1.TextChanged -= RichBoxBody_TextChanged;
-                //流式输出时防闪烁赋值逻辑
-                if (!isStreaming || RichBoxBody.Text != finalTextToShow)
-                {
-                    RichBoxBody.Text = finalTextToShow;
-                }
-                // RichBoxBody.richTextBox1.TextChanged += RichBoxBody_TextChanged;
-				
-				// c. 处理竖排文本（如果需要）
-				if (interface_flag == "从右向左")
-				{
-					RichBoxBody.Text = shupai_Right_txt;
-				}
-				if (interface_flag == "从左向右")
-				{
-					RichBoxBody.Text = shupai_Left_txt;
-				}
-				// (可选) 强制滚动到底部或顶部，有助于解决滚动条问题
-				// RichBoxBody.SelectionStart = 0;
-				// RichBoxBody.ScrollToCaret();
-
-				// 3. 再次显示控件，强制进行一次完整的、干净的重绘
-				RichBoxBody.Visible = true;
-				// 检查是否为截图翻译模式(初版)
-				// if (isScreenshotTranslateMode)
-				// {
-				//     // 检查OCR是否成功获取到文本
-				//     bool success = !string.IsNullOrWhiteSpace(finalTextToShow) && 
-				//                    !finalTextToShow.Contains("***该区域未发现文本***");
-
-				//     if (success)
-				//     {
-				//         // OCR成功，执行核心操作：
-				//         // 1. 将OCR结果放入原文框
-				//         RichBoxBody.Text = finalTextToShow;
-
-				//         // 2. 调用翻译方法，并传入 true 来默认隐藏原文
-				//         TransClick(true); 
-				//     }
-				//     else
-				//     {
-				//         // OCR失败，给用户一个提示，但不显示主窗口
-				//         string errorMessage = string.IsNullOrWhiteSpace(typeset_txt) ? "未识别到文本" : typeset_txt.Replace("***", "").Trim();
-				//         CommonHelper.ShowHelpMsg("截图翻译失败：" + errorMessage);
-				//     }
-
-				//     // 清理并退出，不再执行后续的常规显示逻辑
-				//     HelpWin32.UnregisterHotKey(Handle, 222); 
-				//     StaticValue.IsCapture = false; 
-				//     image_screen?.Dispose();
-				//     return; // 【关键】直接返回，中断后续的标准流程
-				// }
-				// 检查是否为截图翻译模式（包含了显示窗口和不显示窗口两种情况）
-				if (isScreenshotTranslateMode)
-				{
-					bool success = !string.IsNullOrWhiteSpace(finalTextToShow) &&
-								   !finalTextToShow.Contains("***该区域未发现文本***");
-
-					if (success)
-					{
-						// 根据“不显示窗口”的设置决定走哪个流程
-						if (StaticValue.NoWindowScreenshotTranslation)
-						{
-							// 【流程 A：不显示窗口，直接复制（真·静默翻译）】
-							this.Hide();
-							// 或者
-							// this.Visible = false;
-
-							// 异步获取翻译结果
-							string translationResult = await GetTranslationAsync(finalTextToShow);
-
-							// 检查翻译是否成功，这行代码会等到翻译完成后才执行，此时 translationResult 是有值的
-							if (!string.IsNullOrEmpty(translationResult) && !translationResult.Contains("]："))
-							{
-								SetClipboardWithLock(translationResult);
-								CommonHelper.ShowHelpMsg("译文已复制");
-							}
-							else
-							{
-								// 如果翻译出错，也提示用户
-								CommonHelper.ShowHelpMsg("翻译失败：" + translationResult);
-							}
-							// 【关键修正】在此流程的末尾，手动重置标志位
-							isScreenshotTranslateMode = false;
-						}
-						else
-						{
-							// 【流程 B：显示窗口】
-
-							// 1. 将OCR结果放入原文框
-							RichBoxBody.Text = finalTextToShow;
-
-							// 2. 调用翻译方法，并传入 true 来默认隐藏原文
-							TransClick(true);
-						}
-					}
-					else
-					{
-						// OCR失败，不会有翻译流程，所以在这里重置标志是安全的
-						isScreenshotTranslateMode = false;
-						// OCR失败的提示
-						string errorMessage = string.IsNullOrWhiteSpace(typeset_txt) ? "未识别到文本" : typeset_txt.Replace("***", "").Trim();
-						CommonHelper.ShowHelpMsg("截图翻译失败：" + errorMessage);
-					}
-
-					// 统一的清理工作
-					HelpWin32.UnregisterHotKey(Handle, 222);
-					StaticValue.IsCapture = false;
-					image_screen?.Dispose();
-					return; // 中断后续的标准流程
-				}
-
-				// ====================【核心修改结束】====================
-
-				// 原有的 RichBoxBody.Refresh() 不再需要，可以删除
-
-				// --- 步骤 3: 在UI完全稳定后，才执行所有可能阻塞的操作（如剪贴板） ---
-
-				if (shouldPerformCopy)
-				{
-					SetClipboardWithLock(textToCopy);
-					Debug.WriteLine("拆分后自动复制成功");
-
-				}
-				else
-				{
-					//// 处理识别后自动复制功能 (只有同时开启了 ① 识别后自动复制 和 ② 自动翻译 和 ③ 翻译后自动复制，才不复制识别结果)
-					if (autoCopyOcr && (!autoTranslate || !autoCopyTranslate))
-					{
-						SetClipboardWithLock(RichBoxBody.Text);
-						Debug.WriteLine("识别后自动复制成功");
-
-					}
-				}
-
-				// --- 步骤 4: 触发后续逻辑 ---
-				// (百度搜索和无弹窗模式会提前返回)
-				if (baidu_flags == "百度")
-				{
-					FormBorderStyle = FormBorderStyle.Sizable;
-					// Size = new Size((int)font_base.Width * 23, (int)font_base.Height * 24);
-					this.Size = this.lastNormalSize;
-					Visible = false;
-					WindowState = FormWindowState.Minimized;
-					Show();
-					Process.Start("https://www.baidu.com/s?wd=" + RichBoxBody.Text);
-					baidu_flags = "";
-					if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-					{
-						var value = IniHelper.GetValue("快捷键", "翻译文本");
-						var text2 = "None";
-						var text3 = "F9";
-						SetHotkey(text2, text3, value, 205);
-					}
-					HelpWin32.UnregisterHotKey(Handle, 222);
-					return;
-				}
-
-				// 处理识别后自动翻译功能
-				if (autoTranslate)
-				{
-					try
-					{
-						auto_fla = "";
-						isOcrTranslation = true;
-						BeginInvoke(new Translate(TransClick)); // 使用BeginInvoke避免阻塞
-					}
-					catch { }
-				}
-				// 处理文本检查功能
-				if (bool.Parse(IniHelper.GetValue("工具栏", "检查")))
-				{
-					try
-					{
-						RichBoxBody.Find = "";
-					}
-					catch
-					{
-						//
-					}
-				}
-
-				// --- 步骤 5: 最后收尾 ---
-				// 重新设置热键
-				if (IniHelper.GetValue("快捷键", "翻译文本") != "请按下快捷键")
-				{
-					var value3 = IniHelper.GetValue("快捷键", "翻译文本");
-					SetHotkey("None", "F9", value3, 205);
-				}
-				HelpWin32.UnregisterHotKey(Handle, 222);
-			}
-			finally
-			{
-				// ==================== 【状态重置点】 ====================
-
-				// 1. 无论如何，重置流式标记
-				isStreaming = false;
-				
-                if (transtalate_fla == "开启")
-                {
-                    this.RichBoxBody.richTextBox1.TextChanged -= RichBoxBody_TextChanged;
-                    this.RichBoxBody.richTextBox1.TextChanged += RichBoxBody_TextChanged;
-                }
-                else
-                {
-                    this.RichBoxBody.richTextBox1.TextChanged -= RichBoxBody_TextChanged;
-					 // 【新增】防止后台读秒
-            		if (translationTimer != null) translationTimer.Stop(); 
-                }
-
-                // 2. 无论如何，确保文本框解锁 (防止万一报错导致文本框锁死)
-                // 即使是静默模式，解锁一下也无害
-                if (RichBoxBody != null && RichBoxBody.richTextBox1 != null)
-				{
-					//恢复ocr结果框的工具栏
-                    RichBoxBody.SetToolbarEnabled(true);
-                    RichBoxBody.richTextBox1.ReadOnly = false;
-				}
-
-				// 3. 截图翻译模式重置 (如果之前是在 return 前做的，也可以移到这里)
-				// 注意：如果你希望仅在特定情况下重置，保留在上面；
-				// 但通常这种状态标记都适合在这里兜底重置
-				// isScreenshotTranslateMode = false; 
-
-				LogState("Main_OCR_Thread_last End (Cleanup Done)");
-			}
-		}
-
-        private void LogState(string eventName)
-		{
-			// C# 6.0+ string interpolation, simpler
-			System.Diagnostics.Debug.WriteLine($"--- {eventName} ---");
-			System.Diagnostics.Debug.WriteLine($"  Timestamp: {DateTime.Now:HH:mm:ss.fff}");
-			System.Diagnostics.Debug.WriteLine($"  WindowState: {this.WindowState}");
-			System.Diagnostics.Debug.WriteLine($"  Size: {this.Size}");
-			System.Diagnostics.Debug.WriteLine($"  ClientRectangle.Size: {this.ClientRectangle.Size}");
-			System.Diagnostics.Debug.WriteLine($"  transtalate_fla: {this.transtalate_fla ?? "null"}");
-			System.Diagnostics.Debug.WriteLine($"  isOriginalTextHidden: {this.isOriginalTextHidden}");
-			System.Diagnostics.Debug.WriteLine($"  lastNormalSize: {this.lastNormalSize}");
-			System.Diagnostics.Debug.WriteLine("--------------------");
-		}
-        /// <summary>
-        /// 百度OCR中英文识别选项点击事件处理函数
-        /// 设置百度OCR语言为中英文混合识别模式，并更新界面显示
-        /// </summary>
-        /// <param name="sender">事件发送者</param>
-        /// <param name="e">事件参数</param>
-        private void OCR_baidu_Ch_and_En_Click(object sender, EventArgs e)
-		{
-			IniHelper.SetValue("密钥_百度", "language_code", "CHN_ENG");
-			OCR_foreach("中英");
-		}
-
-		/// <summary>
-		/// 百度OCR日语识别选项点击事件处理函数
-		/// 设置百度OCR语言为日语识别模式，并更新界面显示
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_baidu_Jap_Click(object sender, EventArgs e)
-		{
-			IniHelper.SetValue("密钥_百度", "language_code", "JAP");
-			OCR_foreach("日语");
-		}
-
-		/// <summary>
-		/// 百度OCR韩语识别选项点击事件处理函数
-		/// 设置百度OCR语言为韩语识别模式，并更新界面显示
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_baidu_Kor_Click(object sender, EventArgs e)
-		{
-			IniHelper.SetValue("密钥_百度", "language_code", "KOR");
-			OCR_foreach("韩语");
-		}
-
-		/// <summary>
 		/// 检查并处理字符串中的标点符号
 		/// </summary>
 		/// <param name="text">需要处理的文本</param>
@@ -5055,59 +3436,6 @@ namespace TrOCR
 				}
 			}
 			return new string(array);
-		}
-
-		/// <summary>
-		/// 执行搜狗OCR识别功能
-		/// 调用OCR识别接口并处理识别结果，根据设置决定是否分段显示
-		/// </summary>
-		public void SougouOCR()
-		{
-			try
-			{
-				split_txt = "";
-				Image image = ZoomImage((Bitmap)image_screen, 120, 120);
-				//var value = OcrHelper.SgOcr(image);
-				var value = OcrHelper.SgBasicOpenOcr(image);
-				var jArray = JArray.Parse(((JObject)JsonConvert.DeserializeObject(value))["result"].ToString());
-				if (IniHelper.GetValue("工具栏", "分段") == "True")
-				{
-					checked_location_sougou(jArray, 1, "content", "frame");
-				}
-				else
-				{
-					checked_txt(jArray, 1, "content");
-				}
-				image.Dispose();
-			}
-			catch
-			{
-				if (esc != "退出")
-				{
-					RichBoxBody.Text = "***该区域未发现文本***";
-				}
-				else
-				{
-					RichBoxBody.Text = "***该区域未发现文本***";
-					esc = "";
-				}
-			}
-		}
-
-		/// <summary>
-		/// 合并三个字节数组为一个数组
-		/// </summary>
-		/// <param name="a">第一个字节数组</param>
-		/// <param name="b">第二个字节数组</param>
-		/// <param name="c">第三个字节数组</param>
-		/// <returns>合并后的字节数组</returns>
-		public static byte[] MergeByte(byte[] a, byte[] b, byte[] c)
-		{
-			var array = new byte[a.Length + b.Length + c.Length];
-			a.CopyTo(array, 0);
-			b.CopyTo(array, a.Length);
-			c.CopyTo(array, a.Length + b.Length);
-			return array;
 		}
 
 		/// <summary>
@@ -5281,313 +3609,6 @@ namespace TrOCR
 			// 将处理后的文本分别赋值给split_txt和typeset_txt字段
 			split_txt = str + JObject.Parse(jarray[jarray.Count - 1].ToString())[words];
 			typeset_txt = text.Replace("\r\n\r\n", "\r\n") + JObject.Parse(jarray[jarray.Count - 1].ToString())[words];
-		}
-
-		/// <summary>
-		/// 根据名称设置OCR接口类型，并更新相关UI和配置文件
-		/// </summary>
-		/// <param name="name">OCR接口名称</param>
-		private void OCR_foreach(string name)
-		{
-			OcrHelper.Dispose();
-			var filePath = AppDomain.CurrentDomain.BaseDirectory + "Data\\config.ini";
-
-			if (name == "腾讯手写")
-			{
-				
-				MessageBox.Show("请使用腾讯-高精度接口进行手写识别。", "提示", MessageBoxButtons.OK, MessageBoxIcon.Information);
-				return;
-
-			}
-			
-			// --- 【核心判断逻辑：检查是否在冲突引擎间切换】 ---
-			// 1. 定义哪些引擎是互相冲突的
-			var conflictingEngines = new[] { "PaddleOCR", "PaddleOCR2" };
-
-			// 2. 判断：如果 当前接口 和 新接口 都属于冲突列表，并且两者不相同
-			if (interface_flag != name &&
-			    conflictingEngines.Contains(interface_flag) &&
-			    conflictingEngines.Contains(name))
-			{
-				HelpWin32.IniFileHelper.SetValue("配置", "接口", name, filePath);
-
-				// 1. 准备询问对话框的文本内容
-				string message = "PaddleOCR和PaddleOCR2互相切换，需要重启程序才能生效。如果不重启，直接进行识别程序会闪退\n\n是否立即重启？";
-				string caption = "重启应用程序";
-				MessageBoxButtons buttons = MessageBoxButtons.YesNo;
-				DialogResult result;
-
-				// 2. 显示对话框，并获取用户的选择结果
-				result = MessageBox.Show(message, caption, buttons, MessageBoxIcon.Question);
-
-				// 3. 判断用户是否点击了“是”
-				if (result == DialogResult.Yes)
-				{
-					// 4. 如果是，则调用内置的重启方法
-					Application.Restart();
-					// 也可以使用手动重启方案以保证可靠性
-					// System.Diagnostics.Process.Start(Application.ExecutablePath);
-					// Application.Exit();
-				}
-				
-			}
-
-			// 当切换到其他OCR接口时，释放PaddleOCR资源
-			if (interface_flag == "PaddleOCR" && name != "PaddleOCR")
-			{
-				// 如果是从 PaddleOCR 切换到其他接口，就调用 Reset 释放资源
-				try
-				{
-					PaddleOCRHelper.Reset();
-
-				}
-				catch (Exception ex)
-				{
-					// 记录可能发生的释放错误
-					CommonHelper.AddLog($"释放 PaddleOCR 引擎时出错: {ex.Message}");
-				}
-			}
-			
-			// 当切换到其他OCR接口时，释放PaddleOCR2资源
-			if (interface_flag == "PaddleOCR2" && name != "PaddleOCR2")
-			{
-				// 如果是从 PaddleOCR2 切换到其他接口，就调用 Reset 释放资源
-				try
-				{
-					PaddleOCR2Helper.Reset();
-        		}
-				catch (Exception ex)
-				{
-					// 记录可能发生的释放错误
-					CommonHelper.AddLog($"释放 PaddleOCR2 引擎时出错: {ex.Message}");
-				}
-			}
-			
-			// 当切换到其他OCR接口时，释放RapidOCR资源
-			if (interface_flag == "RapidOCR" && name != "RapidOCR")
-			{
-				// 如果是从 RapidOCR 切换到其他接口，就调用 Reset 释放资源
-				try
-				{
-					RapidOCRHelper.Reset();
-        		}
-				catch (Exception ex)
-				{
-					// 记录可能发生的释放错误
-					CommonHelper.AddLog($"释放 RapidOCR 引擎时出错: {ex.Message}");
-				}
-			}
-          
-            // === 【核心修复】: 如果选的不是 AI 接口，强制清除 AI 菜单的状态 ===
-            if (!string.IsNullOrEmpty(name) && name != "CustomOpenAI") // 只要不是点 AI
-            {
-                // 1. 清除 AI 主菜单的勾选和文字
-                if (ai_menu != null)
-                {
-                    //ai_menu.Checked = false;
-                    ai_menu.Text = "AI"; // 恢复默认文字，去掉 "√" 或 "DeepSeek..."
-
-                    // 2. 清除 AI 子菜单的勾选 (可选，保持子菜单选中状态也不错，看你习惯
-					// 遍历所有AI接口(厂商) (第二级)
-                    foreach (ToolStripItem item in ai_menu.DropDownItems)
-                    {
-                        if (item is ToolStripMenuItem providerItem)
-                        {
-                            // 清除厂商勾选
-                            providerItem.Checked = false;
-
-                            // 3.  关键：深入遍历模式 (第三级) 并清除勾选 
-                            if (providerItem.HasDropDownItems)
-                            {
-                                foreach (ToolStripItem subItem in providerItem.DropDownItems)
-                                {
-                                    if (subItem is ToolStripMenuItem modeItem)
-                                    {
-                                        modeItem.Checked = false; // 清除模式勾选
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            switch (name)
-			{
-				case "韩语":
-					interface_flag = "韩语";
-					Refresh();
-					baidu.Text = "百度√";
-					kor.Text = "韩语√";
-					break;
-				case "日语":
-					interface_flag = "日语";
-					Refresh();
-					baidu.Text = "百度√";
-					jap.Text = "日语√";
-					break;
-				case "中英":
-					interface_flag = "中英";
-					Refresh();
-					baidu.Text = "百度√";
-					ch_en.Text = "中英√";
-					break;
-				case "搜狗":
-					interface_flag = "搜狗";
-					Refresh();
-					sougou.Text = "搜狗√";
-					break;
-				case "腾讯":
-					interface_flag = "腾讯";
-					Refresh();
-					tencent.Text = "腾讯√";
-					break;
-				case "腾讯-高精度":
-					interface_flag = "腾讯-高精度";
-					Refresh();
-					tencent_accurate.Text = "腾讯-高精度√";
-					break;
-				case "有道":
-					interface_flag = "有道";
-					Refresh();
-					youdao.Text = "有道√";
-					break;
-				case "微信":
-					interface_flag = "微信";
-					Refresh();
-					wechat.Text = "微信√";
-					break;
-				case "白描":
-					interface_flag = "白描";
-					Refresh();
-					baimiao.Text = "白描√";
-					break;
-				case "百度-高精度":
-					interface_flag = "百度-高精度";
-					Refresh();
-					baidu_accurate.Text = "百度-高精度√";
-					break;
-				 case "百度手写":
-        			interface_flag = "百度手写";
-        			Refresh(); // 这个方法会重置所有√
-        			write.Text = "手写√"; // 将"手写"按钮标记为选中
-					baidu_handwriting.Text = "百度手写√";
-        			break;		
-				case "公式":
-					interface_flag = "公式";
-					Refresh();
-					Mathfuntion.Text = "公式√";
-					break;
-				case "百度表格":
-					interface_flag = "百度表格";
-					Refresh();
-					ocr_table.Text = "表格√";
-					baidu_table.Text = "百度√";
-					break;
-				case "腾讯表格":
-					interface_flag = "腾讯表格";
-					Refresh();
-					ocr_table.Text = "表格√";
-					tx_table.Text = "腾讯√";
-				break;
-				case "阿里表格":
-					interface_flag = "阿里表格";
-					Refresh();
-					ocr_table.Text = "表格√";
-					ali_table.Text = "阿里√";
-					break;
-				case "PaddleOCR":
-					interface_flag = "PaddleOCR";
-					Refresh();
-					paddleocr.Text = "PaddleOCR√";
-					break;
-				case "PaddleOCR2":
-					interface_flag = "PaddleOCR2";
-					Refresh();
-					paddleocr2.Text = "PaddleOCR2√";
-					break;
-				case "RapidOCR":
-					interface_flag = "RapidOCR";
-					Refresh();
-					rapidocr.Text = "RapidOCR√";
-					break;
-                case "CustomOpenAI":
-                    interface_flag = "CustomOpenAI";
-                    Refresh(); // 先重置所有菜单文字
-                               // 这里先设置一个基础状态，具体的 "AI: DeepSeek..." 文字
-                               // 会由 SwitchToCustomAI 方法在后面覆盖更新
-                    ai_menu.Text = "AI√";
-                    break;
-                case "从左向右" when !File.Exists("cvextern.dll"):
-					MessageBox.Show("请从蓝奏网盘中下载cvextern.dll大小约25m，点击确定自动弹出网页。\r\n将下载后的文件与 天若OCR文字识别.exe 这个文件放在一起。");
-					Process.Start("https://www.lanzous.com/i1ab3vg");
-					break;
-				case "从左向右":
-					interface_flag = "从左向右";
-					Refresh();
-					shupai.Text = "竖排√";
-					left_right.Text = "从左向右√";
-					break;
-				case "从右向左" when !File.Exists("cvextern.dll"):
-					MessageBox.Show("请从蓝奏网盘中下载cvextern.dll大小约25m，点击确定自动弹出网页。\r\n将下载后的文件与 天若OCR文字识别.exe 这个文件放在一起。");
-					Process.Start("https://www.lanzous.com/i1ab3vg");
-					return;
-				case "从右向左":
-					interface_flag = "从右向左";
-					Refresh();
-					shupai.Text = "竖排√";
-					righ_left.Text = "从右向左√";
-					break;
-			}
-
-			HelpWin32.IniFileHelper.SetValue("配置", "接口", interface_flag, filePath);
-		}
-
-		/// <summary>
-		/// OCR识别方向设置为竖排的事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_shupai_Click(object sender, EventArgs e)
-		{
-		}
-
-		/// <summary>
-		/// OCR识别设置为手写的事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		// private void OCR_write_Click(object sender, EventArgs e)
-		// {
-		// 	OCR_foreach("手写");
-		// }
-		private void OCR_baidu_handwriting_Click (object sender, EventArgs e)
-		{
-			OCR_foreach("百度手写");
-		}
-		private void OCR_tencent_handwriting_Click (object sender, EventArgs e)
-		{
-			OCR_foreach("腾讯手写");
-		}
-
-		/// <summary>
-		/// OCR识别方向设置为从左向右的事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_lefttoright_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("从左向右");
-		}
-
-		/// <summary>
-		/// OCR识别方向设置为从右向左的事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_righttoleft_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("从右向左");
 		}
 
 		/// <summary>
@@ -6996,27 +5017,6 @@ namespace TrOCR
 				return $"[腾讯翻译]：发生未知错误 - {ex.Message}";
 			}
 		}
-		/// <summary>
-		/// 百度手写识别
-		/// </summary>
-		public void OCR_baidu_handwriting()
-		{
-		    split_txt = "";
-			try
-			{
-				var imageBytes = OcrHelper.ImgToBytes(image_screen);
-				// 【修改】将配置的语言传递给识别方法
-				var result = BaiduOcrHelper.Handwriting(imageBytes, StaticValue.BD_HANDWRITING_LANGUAGE);
-				// ProcessOcrResult(result); //不需要再次处理
-				split_txt = typeset_txt = result;
-				
-		    }
-			catch (Exception ex)
-			{
-				typeset_txt = $"***百度手写识别出错: {ex.Message}***";
-				split_txt = typeset_txt;
-			}
-		}
 		private void ExportExcel_Click(object sender, EventArgs e)
 		{
 		    // 调用我们创建的帮助类来处理导出
@@ -7434,65 +5434,6 @@ namespace TrOCR
 			dataObject.SetData(DataFormats.UnicodeText, data);
 			SetClipboardWithLock(dataObject);
 			Debug.WriteLine("识别表格结果写入剪贴板");
-		}
-
-		/// <summary>
-		/// OCR数学公式识别点击事件处理函数
-		/// </summary>
-		/// <param name="sender">事件发送者</param>
-		/// <param name="e">事件参数</param>
-		private void OCR_Mathfuntion_Click(object sender, EventArgs e)
-		{
-			OCR_foreach("公式");
-		}
-
-		/// <summary>
-		/// 使用OCR技术识别图像中的数学公式
-		/// </summary>
-		public void OCR_Math()
-		{
-			split_txt = "";
-			try
-			{
-				var img = image_screen;
-				var inArray = OcrHelper.ImgToBytes(img);
-				// 构造发送到Mathpix API的JSON数据
-				var s = "{\t\"formats\": [\"latex_styled\", \"text\"],\t\"metadata\": {\t\t\"count\": 0,\t\t\"platform\": \"windows 10\",\t\t\"skip_recrop\": true,\t\t\"user_id\": \"\",\t\t\"version\": \"snip.windows@01.02.0027\"\t},\t\"ocr\": [\"text\", \"math\"],\t\"src\": \"data:image/jpeg;base64," + Convert.ToBase64String(inArray) + "\"}";
-				var bytes = Encoding.UTF8.GetBytes(s);
-				// 创建并配置HTTP请求
-				var httpWebRequest = (HttpWebRequest)WebRequest.Create("https://api.mathpix.com/v3/latex");
-				httpWebRequest.Method = "POST";
-				httpWebRequest.ContentType = "application/json";
-				httpWebRequest.Timeout = 8000;
-				httpWebRequest.ReadWriteTimeout = 5000;
-				httpWebRequest.Headers.Add("app_id: mathpix_chrome");
-				httpWebRequest.Headers.Add("app_key: 85948264c5d443573286752fbe8df361");
-				using (var requestStream = httpWebRequest.GetRequestStream())
-				{
-					requestStream.Write(bytes, 0, bytes.Length);
-				}
-				// 发送请求并获取响应
-				var responseStream = ((HttpWebResponse)httpWebRequest.GetResponse()).GetResponseStream();
-				var value = new StreamReader(responseStream, Encoding.GetEncoding("utf-8")).ReadToEnd();
-				responseStream.Close();
-				// 解析响应结果，提取LaTeX格式的数学公式
-				var text = "$" + ((JObject)JsonConvert.DeserializeObject(value))["latex_styled"] + "$";
-				split_txt = text;
-				typeset_txt = text;
-			}
-			catch
-			{
-				// 处理异常情况并显示相应错误信息
-				if (esc != "退出")
-				{
-					RichBoxBody.Text = "***该区域未发现文本或者密钥次数用尽***";
-				}
-				else
-				{
-					RichBoxBody.Text = "***该区域未发现文本***";
-					esc = "";
-				}
-			}
 		}
 
 #endregion
